refactor: fix type errors

+++ packages/backend/src/graphql/mutations/delete-step.ts
@@ -19,6 +19,8 @@ const deleteStep = async (
     })
     .throwIfNotFound();
 
+  if (!step) return;
+
   await step.$query().delete();
 
   const nextSteps = await step.flow
